Update Precision and Blance to float for addressAssetBalance. Duplicate transactions hash fix in blockpersister

+++ StateOfNeo/StateOfNeo.Data.Models/AddressAssetBalance.cs
@@ -8,7 +8,7 @@ namespace StateOfNeo.Data.Models
     {
         public int Id { get; set; }
 
-        public decimal Balance { get; set; }
+        public float Balance { get; set; }
 
         public int TransactionsCount { get; set; }
